pyproject.toml
[project]
name = "fluxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "faker>=37.8.0",
    "fastapi>=0.116.1",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "google-genai>=1.26.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "uvloop>=0.21.0",
    "xlsxwriter>=3.2.5",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",
    "faker>=20.0.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=80",
]
markers = [
    "unit: tests unitaires",
    "integration: tests d'intégration",
    "slow: tests lents",
    "database: tests nécessitant une base de données",
]
